[Mono.Android] Build latest stable version against netcoreapp3.1 (#4419)
We expect the .NET 5 BCL to closely resemble the .NET Core 3.1 BCL.
Continue preparing for .NET 5 and build `Mono.Android.dll` against
the `netcoreapp3.1` target framework, to ensure that the APIs that we
need actually exist.

Additionally, the `Mono.Android.csproj` build will effectively be a
no-op when targeting `netcoreapp3.1` if `$(AndroidFrameworkVersion)`
is not equal to `$(AndroidLatestStableFrameworkVersion)`, as we
currently plan on only shipping *one* `Mono.Android.dll` assembly as
part of .NET 5, so there is no need to build `Mono.Android.dll`
against other `$(TargetFrameworkVersion)` values.

When building for `netcoreapp3.1`, `Mono.Android.dll` uses the
`Java.Interop.dll` which is built for .NET Standard, instead of the
one built against the `MonoAndroid` framework.  We may want to
consider updating `Java.Interop.csproj` to multitarget against
`netcoreapp3.1` as well.