[Xamarin.Android.Build.Tasks] Run `aapt2 compile` incrementally (#4190)
The `aapt2 compile` command runs in two modes.  The one we currently
use is the `archive` mode: we call `aapt2 compile` with the `--dir`
argument tp gemerate one `.flata` archive for all the files.  The
side effect of this is that even if you only change one file, it will
need to regenerate the entire `.flata` archive.

There is a second mode: rather than using `--dir` you just send in a
single file.  This then writes a single `.flat` file to the output
directory.  While this does mean you have to call `aapt2 compile` for
*every* file, it does mean we can leverage MSBuild's support for
partial targets.  This means MSBuild will detect *only* the files
which changed and allow us to call `aapt2 compile` with just the
modified files.

One exception to this new system are references which use the
`%(AndroidSkipResourceProcessing)` metadata.  In those cases the
chance of those libraries being updated on a regular basis is quite
low, so in that case using a `flata` archive will be better since the
files won't be changing often.

While this may impact initial build times, the goal is to make
incremental builds quicker.  This is especially true for users which
use *lots* of `@(AndroidResource)` items.

A note regarding the `aapt2 daemon` mode: in order to write accented
characters we need to set the `Process.StandardInput` encoding to
UTF-8.  This is not possible directly in .NET Standard 2.0.  We have
to instead use `Console.InputEncoding`.  Also note that we *cannot*
include a BOM when writing the commands.  This is because `aapt2` will
try to parse the BOM as command characters.

Finally, the `aapt2 link` command sometimes reports it is "Done"
before it has even written the archive for the file.  We can get into
a position where we think we are done but the file is not on disk.
Work around this by including a nasty wait which will poll for the
existence of the expected output file and only return when it exists.
The good news is we know at this point if the command failed or not,
so we can bypass the check on failure.

By using incremental builds, we improve incremental build times:

| Test         | Latest Stable | Aapt2 Daemon |
|--------------|---------------|--------------|
| From Clean   | 00:00:27.92   | 00:00:27.41  |
| No Changes   | 00:00:03.14   | 00:00:03.72  |
| Alter Layout | 00:00:16.37   | 00:00:05.53  |

Note that **Alter Layout** with Aapt2 Daemon runs in 34% of the time.