[Xamarin.Android.Build.Tasks] clean $(IntermediateOutputPath) if build.props changes (#4360)
I found a crash at runtime, if you:

 1. Build / Run a project with 16.4
 2. Build / Run the same project with 16.6

The stack trace is:

	java.lang.UnsatisfiedLinkError: dlopen failed: cannot locate symbol "jm_typemap" referenced by "/data/app/com.companyname.app59-bgIEcmaPpoMxpR948Wo8ZA==/lib/x86/libmonodroid.so"...
	    at java.lang.Runtime.loadLibrary0(Runtime.java:1071)
	    at java.lang.Runtime.loadLibrary0(Runtime.java:1007)
	    at java.lang.System.loadLibrary(System.java:1667)
	    at mono.MonoPackageManager.LoadApplication(MonoPackageManager.java:80)
	    at mono.MonoRuntimeProvider.attachInfo(MonoRuntimeProvider.java:48)
	    at android.app.ActivityThread.installProvider(ActivityThread.java:6983)
	    …

What happened is in commit ce2bc689 the `jm_typemap` and `mj_typemap`
symbols were removed.  But we don't actually have anything that
*cleans* the generated assembly files when `$(XamarinAndroidVersion)`
changes, so there were some "stale" files from the 16.4 build which
referenced symbols that had been removed in 16.6.

We have a `DeleteBinObjTest` test fixture, why didn't that catch this?

Apparently the test was triggering:

	Building target "_CleanIntermediateIfNuGetsChange" completely.
	Input file "HelloForms.Android\obj\project.assets.json" is newer than output file "obj\Debug\90\stamp\_CleanIntermediateIfNuGetsChange.stamp".

And so the test wasn't actually *testing* much of anything!

As part of that test, we have to run a NuGet restore.  The zip file
contains a project that was restored on another machine.  We *have*
to run `/t:Restore` for things to build, but that `Restore` prevented
us from noticing this particular execution path.

To fix the real problem here, I think we need to fundamentally change
what happens when values in `build.props` /
`$(_AndroidBuildPropertiesCache)` changes.  Currently we are using
`build.props` invalidation" to trigger several MSBuild targets, but I
believe we should also use it to trigger the
`_CleanMonoAndroidIntermediateDir` target when `build.props` changes:

 1. Repurpose `_CleanIntermediateIfNuGetsChange` to instead be a
    `_CleanIntermediateIfNeeded` target that does not run on
    design-time builds.

 2. Include the `File.GetLastWriteTime()` value of
   `$(_NuGetAssetsFile)` in `build.props`.

 3. Remove `DesignTimeBuild=$(DesignTimeBuild)` from `build.props`.
    We use a different path for Design-Time-Builds at
    `obj\Debug\designtime\build.props` anyway.

 4. The `_CleanMonoAndroidIntermediateDir` target can't delete
    `build.props` anymore.  I moved this so it will still happen
    during a full `Clean`.

 5. Remove the `_CleanIntermediateIfPackageNamingPolicyChanges` and
    `_CleanupOldStaticResources` MSBuild targets.  They are no longer
    needed at all.

 6. Remove `Xamarin.Android.DefaultOutputPaths.targets`.


I also updated the `DeleteBinObjTest` test to include projects built
with 16.4 and verify the app launches.