[tests] ignore failure to delete MSBuild test directories
Context: https://build.azdo.io/3594621

We have commonly seen test failures on CI, such as:

    Xamarin.Android.Build.Tests.ManifestTest.VersionCodeTests

    TearDown : System.IO.IOException : The directory is not empty.
        --TearDown
        at System.IO.Directory.DeleteHelper(String fullPath, String userPath, Boolean recursive, Boolean throwOnTopLevelDirectoryNotFound, WIN32_FIND_DATA& data)
        at System.IO.Directory.Delete(String fullPath, String userPath, Boolean recursive, Boolean checkHost)
        at Xamarin.Android.Build.Tests.BaseTest.CleanupTest() in /Users/builder/azdo/_work/94/s/xamarin-android/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/BaseTest.cs:line 443

I was not able to find a process running that seems like it would hold
a lock on a file in a sub-directory. It also doesn't seem like a
readonly file could cause this:

    FileSystemUtils.SetDirectoryWriteable (output);
    Directory.Delete (output, recursive: true);

For now, I think we should just change a failure like this so it
doesn't fail the test.

Another failure we've seen is:

    Xamarin.Android.Build.Tests.AndroidUpdateResourcesTest.CheckMaxResWarningIsEmittedAsAWarning(True)
    (_ConvertResourcesCases target) ->
        Resources\values-v27\Strings.xml : warning XA1001: AndroidResgen: Warning while updating resource XML 'obj\Release\80\res\values-v27\strings.xml': Access to the path 'C:\A\vs2019xam000008-1\_work\2\s\bin\TestRelease\temp\CheckMaxResWarningIsEmittedAsAWarningTrue\obj\Release\80\res\values-v27\strings.xml' is denied.
    (_CompileResources target) ->
        Resources\values-v27\Strings.xml : error APT2126: file not found.

`XA1001` loses the stack trace because it is only printing
`Exception.Message`, so I changed this to just do
`Exception.ToString()`. Future builds should give us a stack trace.