[ci] Parallelize MSBuild Test Suites (#4445)
Parallelize our MSBuild test suite (`Xamarin.Android.Build.Tests.dll`)
to run on 3 machines concurrently instead of 1.  This is accomplished
by applying `[Category]` attributes to the NUnit TestFixtures to
specify which "node" they should run on:

	[Category ("Node-3")]
	public class BuildAssetsTest : BaseTest

I've done some initial load balancing between the 3 nodes, but `Node-1`
only runs `class BuildTest` and it needs to be broken down into
multiple classes so we can move some to `Node-2`.  This will be a
future request for @dellis1972 or @jonathanpeppers so we can get a
better distribution.

If a new TestFixture is created and it does not specify a `Node-X`
category it will be run on `Node-2` in the
`run Xamarin.Android.Build.Tests - Mac/Windows - No Node` step.
This is a backup mechanism, it is expected that every TestFixture
specifies a Node.

~~ Results ~~

Previously:

  - MSBuild - MacOS: 1h 43m 48s
  - MSBuild - Windows: 1h 34m 45s

Now:

  - MSBuild - Mac-1: 1h 5m 45s
  - MSBuild - Mac-2: 26m 0s
  - MSBuild - Mac-3: 44m 50s
  - MSBuild - Windows-1: 52m 41s
  - MSBuild - Windows-2: 16m 18s
  - MSBuild - Windows-3: 39m 55s