[Xamarin.Android.Build.Tasks] fix for API R and <GetJavaPlatformJar/> (#4460)
One of our QA tests was failing when using
`TargetFrameworkVersion=v10.0.99`:

	C:\Program Files\Android\jdk\microsoft_dist_openjdk_1.8.0.25\bin\javac.exe -J-Dfile.encoding=UTF8 obj\Release\10099\android\src\mono\MonoRuntimeProvider.java:25: error: cannot find symbol (TaskId:163)
	            String[] splitApks = applicationInfo.splitPublicSourceDirs; (TaskId:163)
	                                                ^ (TaskId:163)
	    symbol:   variable splitPublicSourceDirs (TaskId:163)
	    location: variable applicationInfo of type ApplicationInfo (TaskId:163)
	Note: obj\Release\10099\android\src\crc64e49dab49335e9709\ButtonActivity.java uses unchecked or unsafe operations. (TaskId:163)
	Note: Recompile with -Xlint:unchecked for details. (TaskId:163)

`splitPublicSourceDirs` is available starting with API-21;
what is going on?

	Task "GetJavaPlatformJar" (TaskId:17)
	  Task Parameter:AndroidSdkPlatform=30 (TaskId:17)
	  Task Parameter:AndroidManifest=Properties\AndroidManifest.xml (TaskId:17)
	  Output Property: JavaPlatformJarPath=C:\Program Files %28x86%29\Android\android-sdk\platforms\android-19\android.jar (TaskId:17)
	  Output Property: _AndroidTargetSdkVersion=19 (TaskId:17)

The `AndroidManifest.xml` in this project has:

	<uses-sdk android:targetSdkVersion="19" />

But this could wouldn't be able to parse an API level of R:

	string targetFrameworkVersion = MonoAndroidHelper.SupportedVersions.GetIdFromApiLevel (AndroidSdkPlatform);
	string targetSdkVersion       = MonoAndroidHelper.SupportedVersions.GetIdFromApiLevel (target);

`AndroidApiInfo.xml` for API R right now is:

	<AndroidApiInfo>
	  <Id>R</Id>
	  <Level>30</Level>
	  <Name>R</Name>
	  <Version>v10.0.99</Version>
	  <Stable>False</Stable>
	</AndroidApiInfo>

Rework `<GetJavaPlatformJar/>` to accept an API name of R, and emit
the `XA4211` warning.

I was able to reproduce this issue in a test, but I needed a small fix
in `GetTargetFrameworkVersionRange()`.  It was returning `v10.0` for
`v10.0.99`, due to only using the first two version parts.

I also felt that the `CheckTheCorrectRuntimeAssemblyIsUsedFromNuget()`
test could just hardcode `monoandroid10.0`, as that will continue to
work for new API levels.  I could not get `monoandroid10.0.99` to work.
The test is testing the `monoandroid10.0` assembly is used over
`netstandard2.0`.