[build] /restore no longer needed on Windows
On macOS,  you can generally build with just:

    make prepare all

On Windows:

    msbuild Xamarin.Android.sln /t:Prepare
    msbuild Xamarin.Android.sln

In 295aff28, we introduced some changes that now require a change to
the second step:

    msbuild Xamarin.Android.sln /restore

I think we can add a call at the end of `/t:Prepare`:

    <MSBuild Projects="$(_TopDir)\Xamarin.Android.sln" Targets="Restore" />

This way you don't have to remember the `/restore` switch.

Lastly, I updated the Windows `Build and Smoke Test` phase on Azure
DevOps, so it will build on Windows without `/restore`.