[Xamarin.Android.Build.Tasks] -ignorewarnings option to make r8 less strict
Context: https://www.guardsquare.com/products/proguard/manual/usage
Fixes: https://feedback.devdiv.io/959121
Fixes: https://feedback.devdiv.io/960245
Fixes: https://github.com/xamarin/xamarin-android/issues/4457

A common problem Xamarin.Android developers are running into is `r8`
failing with:

    Warning: Missing class: org.apache.http.client.methods.HttpEntityEnclosingRequestBase
    Error: Compilation can't be completed because some library classes are missing.

The simplest way to hit this is:

  * Use Google Play Services
  * Use multi-dex

In this case, we are only using `r8` for multi-dex support and *not*
code shrinking. Developers aren't setting proguard rules at all in
this scenario.

The correct fix is to do this in `AndroidManifest.xml`:

      <uses-library android:name="org.apache.http.legacy" android:required="false" />
    </application>

This will automatically include `--lib` to
`android-sdk\platforms\android-29\optional\org.apache.http.legacy.jar`
in the call to `r8`.

This is just completely different behavior than you got using `dx`.
Even though the fix is *correct*, it's a problem developers don't want
to run into. Most of the time everything will work fine without doing
this.

This error message can surface in other forms beyond usage of
`org.apache.http.legacy`. One example: A NuGet package has a
dependency on Google Play Services, but doesn't actually include the
`<dependency/>` in its `.nuspec` file. The fix is to add a
`<PackageReference/>` in the project yourself.

What is really weird about this, is `r8` only fails on *some*
warnings, not all.

`r8` is completely fine to emit warnings like these and continue on:

    Warning: The rule `-keep public class * extends java.lang.annotation.Annotation {
    *;
    }` uses extends but actually matches implements.
    Warning in obj\Debug\lp\44\jl\classes.jar:com/google/android/gms/internal/zzam.class:
    Type `org.apache.http.impl.cookie.DateUtils` was not found, it is required for default or static interface methods desugaring of `long com.google.android.gms.internal.zzam.zzf(java.lang.String)`
    Warning in obj\Debug\lp\44\jl\classes.jar:com/google/android/gms/internal/zzas.class:
    Type `android.net.http.AndroidHttpClient` was not found, it is required for default or static interface methods desugaring of `com.google.android.gms.internal.zzs com.google.android.gms.internal.zzas.zza(android.content.Context, com.google.android.gms.internal.zzan)`
    Warning: Resource 'META-INF/MANIFEST.MF' already exists.

I found a configuration option for proguard rules that appears to do
what we want:

    -ignorewarnings

    Specifies to print any warnings about unresolved references and other
    important problems, but to continue processing in any case. Ignoring
    warnings can be dangerous. For instance, if the unresolved classes or
    class members are indeed required for processing, the processed code
    will not function properly. Only use this option if you know what
    you're doing!

Even though there is a scary message, this seems to do exactly what we
need:

    1. r8 continues with its job if there are warnings.
    2. We still actually *emit* the warnings.

The name `-ignorewarnings` is somewhat confusing as it does not
suppress them at all. It just prevents an overall failure.

I added a new `$(AndroidR8IgnoreWarnings)` (`True` by default) MSBuild
property that will automatically include the `-ignorewarnings` option
in the ProGuard configuration file Xamarin.Android generates.

I updated a test for this scenario.