[Xamarin.Android.Build.Tasks] Remove unused error from AndroidApkSigner
Thanks to b28a93ad223f3610067aef63c427afb17a23e7d7, the error check in
`AndroidApkSigner.RunTask()` is no longer needed.  Before that commit,
`RunTask()` was validating the path of `apksigner` or `apksigner.bat`.
But now that `AndroidApkSigner` inherits from `JavaToolTask`, the check
is just revalidating the path of `java`.

One option would be to update the check to validate the path of
`apksigner.jar` instead, but that isn't needed because
`<ResolveAndroidTooling/>` already validates that path.  So the check
can just be removed.