[Xamarin.Android.Build.Tasks] Remove unused error in AndroidApkSigner (#4524)
Thanks to commit b28a93ad, the error check in
`AndroidApkSigner.RunTask()` is no longer needed.  Before b28a93ad,
`RunTask()` was validating the path of `apksigner` or `apksigner.bat`.
Now that `AndroidApkSigner` inherits from `JavaToolTask`, the check is
just revalidating the path of `java`.

One option would be to update the check to validate the path of
`apksigner.jar` instead, but that isn't needed because
`<ResolveAndroidTooling/>` already validates that path.

Remove the path check entirely.