[build] remove remaining packages.config usage
In an effort to port the remaining projects in xamarin-android to
short-form MSBuild projects, a first step is make sure we are using
`<PackageReference/>` everywhere.

I went ahead and migrated a few "easy" projects to short-form:

* `api-merge.csproj`
* `jnienv-gen.csproj`
* `jit-times.csproj`

As a result of these changes `xaprepare` no longer needs to run:

    .nuget\NuGet.exe restore Xamarin.Android.sln

against any of the `.sln` files in this repository. We can rely on
`msbuild /restore` to build and restore everything now.

We still need `.nuget\NuGet.exe`, as it will be used during MSBuild
tests using `packages.config`.