[Xamarin.Android.Build.Tasks] Layout code-behind l10n support
Context: https://github.com/xamarin/xamarin-android/issues/1560
Context: https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1009374/

Add codes to the existing warnings and errors in the layout code-behind
tasks, and move the message strings into the `.resx` file so that they
are localizable.

Remove `<GenerateCodeBehindForLayout/>` so that it doesn't need to be
localized.  The original uses of that task were removed in commit
ab3773cf.