[Xamarin.Android.Build.Tasks] fixup invalid $(AndroidDexTool)
Fixes: https://github.com/xamarin/xamarin-android/issues/4439

If you build a project with `AndroidDexTool=invalid`, you end up with
an APK file that is missing a `classes.dex`. We don't emit an error,
and you won't actually know anything is wrong until going to install
the APK.

If someone did this on accident, I think we can just default
`AndroidDexTool=d8` *unless* it is set to `dx`.

I updated a test for this.