[build] [Mono.Android] Assign $(JavaCallableWrapperAbsAssembly) after $(OutputPath)
I noticed that `GenerateJavaCallableWrappers` was building completely on
every incremental build for me locally:

	Target "GenerateJavaCallableWrappers: (TargetId:1260)" in file "C:\Source\xamarin-android\build-tools\scripts\JavaCallableWrappers.targets" from project "C:\Source\xamarin-android\src\Mono.Android\Mono.Android.csproj" (target "Build" depends on it):
	Building target "GenerateJavaCallableWrappers" completely.
	Input file "C:\Source\xamarin-android\Mono.Android.dll" does not exist.

The issue was that the refactoring to short-form project style in commit
a6e5d10a by chance moved the assignment of
`$(JavaCallableWrapperAbsAssembly)` to happen before the assignment of
`$(OutputPath)`, meaning that `$(JavaCallableWrapperAbsAssembly)` no
longer contained the correct path to the `Mono.Android.dll` output file.

Move the assignment of `$(JavaCallableWrapperAbsAssembly)` back after
`$(OutputPath)` to take care of that.