[docs] Put newest download locations first in README
When the xamarin-macios team recently added a downloads section to their
README, they noticed that users were by default picking the first link,
so they decided to switch the order to sort from newest to oldest.

Update the download locations in the xamarin-android README to use
newest-to-oldest order like the xamarin-macios README, and group them
into "Current" and "Previous" sections.

Put the OSS master download link at the bottom of the "Current" section
because it is expected to be used less frequently than the links for the
commercial versions.

Other changes:

Update the "Status" heading to "Link" because the only remaining
versions where that column contains a status are the old 9.0 and 9.1
versions.