[.NET 5] exclude shared runtime APKs from Xamarin.Android.Sdk
Using `AndroidUseSharedRuntime=True` doesn't work in .NET 5 yet, and
it is still up in the air if we will support it.

For now, we can remove these from `Xamarin.Android.Sdk.nupkg` to save
on file size. This save ~164MB of file size on disk when the NuGet
package is extracted.