Fix instant run issue
The previous approach broke instant run, as it didn't account for all
the ways in which `@(AndroidNativeLibrary)` was used in the monodroid
repo.  A new ItemGroup has been added specifically for the files which
were previously hard coded in the <BuildApk/> task to ensure that they
are processed separately and included in the .apk.