[build] Use API-21 for all native builds
Up until this commit, we have been building the 64-bit targets (`x86-64`
and `armv8-a` against NDK API 21 while the 32-bit targets (`x86` and
`armv7-a`) against NDK API 16.  This commit changes it so that all
targets build against API 21.  This gives us the ability to speed up
Debug builds on 32-bit platforms via use of  POSIX APIs for directory
traversal that are available starting from API 21 onwards (`openat` and
`fstatat`).  The switch to API 21 was initially planned for .NET5 but
there's no harm in making the change now.

In addition, this commit updates versions of several components:

  * Bump emulator version to 30.0.5
  * Bump NDK version to 21b (21.1.6352462)
  * Bump x86 system image revision to 7

Also, `XABuildConfig.NDKMinimumApiAvailable` is now set to the minimum
supported NDK API level instead of the minimum API level *installed* by
the NDK.