[Xamarin.Android.Build.Tasks] JarToXml & BindingsGenerator l10n support
Add codes for the existing errors in `<JarToXml/>` and
`<BindingsGenerator/>`, and move the message strings into the `.resx`
file so that they are localizable.

Other changes:

Adjust the wording of the messages to standardize some terms,
capitalizations, and punctuation.

Remove the `subcategory` parameters from the `LogError()` calls in
`<BindingsGenerator/>` because no other errors in Xamarin.Android use
the `subcategory` parameter of `LogError()`.